requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-requestlogs==0.8
djangorestframework==3.15.2
python-dotenv==1.0.1
setuptools==75.3.0
sqlparse==0.5.1
